Refactor specific weapon classes into dedicated files in dlls/weapons/.
Extracted CRpg, CShotgun, CGauss, CEgon, CSatchel and base weapon classes (CBasePlayerItem, CBasePlayerWeapon, CWeaponBox, CBasePlayerAmmo) into their respective files in the dlls/weapons/ directory.
Updated dlls/weapons.h, dlls/weapons.cpp, Makefiles, and VC++ configuration files to match the new file structure.
Fixed minor cross-platform compilation issues and added necessary `#ifndef CLIENT_DLL` blocks for TYPEDESCRIPTION data.

+++ dlls/core/util.cpp
@@ -23,6 +23,7 @@
 #include "core/extdll.h"
 #include "core/util.h"
 #include "core/cbase.h"
+#include <stdint.h>
 #include "core/saverestore.h"
 #include <time.h>
 #include "shake.h"
@@ -1977,7 +1978,7 @@ void CSave :: WriteFunction( const char *pname, void **data, int count )
 {
 	const char *functionName;
 
-	functionName = NAME_FOR_FUNCTION( (uint32)*data );
+	functionName = NAME_FOR_FUNCTION( (uintptr_t)*data );
 	if ( functionName )
 		BufferField( pname, strlen(functionName) + 1, functionName );
 	else
